[Feature] Modify the code implementation
