+++ flink-doris-connector/src/main/java/org/apache/doris/flink/serialization/RowBatch.java
@@ -32,6 +32,7 @@ import org.apache.arrow.vector.Float8Vector;
 import org.apache.arrow.vector.IntVector;
 import org.apache.arrow.vector.SmallIntVector;
 import org.apache.arrow.vector.TimeStampMicroVector;
+import org.apache.arrow.vector.TimeStampSecVector;
 import org.apache.arrow.vector.TinyIntVector;
 import org.apache.arrow.vector.UInt4Vector;
 import org.apache.arrow.vector.VarBinaryVector;
@@ -41,11 +42,13 @@ import org.apache.arrow.vector.complex.ListVector;
 import org.apache.arrow.vector.complex.MapVector;
 import org.apache.arrow.vector.complex.StructVector;
 import org.apache.arrow.vector.complex.impl.UnionMapReader;
+import org.apache.arrow.vector.ipc.ArrowReader;
 import org.apache.arrow.vector.ipc.ArrowStreamReader;
 import org.apache.arrow.vector.types.Types;
 import org.apache.doris.flink.exception.DorisException;
 import org.apache.doris.flink.exception.DorisRuntimeException;
 import org.apache.doris.flink.rest.models.Schema;
+import org.apache.doris.flink.util.FastDateUtil;
 import org.apache.doris.flink.util.IPUtils;
 import org.apache.doris.sdk.thrift.TScanBatchResult;
 import org.slf4j.Logger;
@@ -55,6 +58,7 @@ import java.io.ByteArrayInputStream;
 import java.io.IOException;
 import java.math.BigDecimal;
 import java.math.BigInteger;
+import java.nio.charset.StandardCharsets;
 import java.time.Instant;
 import java.time.LocalDate;
 import java.time.LocalDateTime;
@@ -93,7 +97,7 @@ public class RowBatch {
     private int rowCountInOneBatch = 0;
     private int readRowCount = 0;
     private final List<Row> rowBatch = new ArrayList<>();
-    private final ArrowStreamReader arrowStreamReader;
+    private final ArrowReader arrowStreamReader;
     private VectorSchemaRoot root;
     private List<FieldVector> fieldVectors;
     private RootAllocator rootAllocator;
@@ -116,6 +120,12 @@ public class RowBatch {
         this.arrowStreamReader =
                 new ArrowStreamReader(
                         new ByteArrayInputStream(nextResult.getRows()), rootAllocator);
+        this.offsetInRowBatch = 0;
+    }
+
+    public RowBatch(ArrowReader nextResult, Schema schema) {
+        this.schema = schema;
+        this.arrowStreamReader = nextResult;
         this.offsetInRowBatch = 0;
     }
 
@@ -150,6 +160,37 @@ public class RowBatch {
             throw new DorisRuntimeException(e.getMessage());
         } finally {
             close();
+        }
+    }
+
+    public RowBatch readFlightArrow() {
+        try {
+            this.root = arrowStreamReader.getVectorSchemaRoot();
+            fieldVectors = root.getFieldVectors();
+            if (fieldVectors.size() > schema.size()) {
+                logger.error(
+                        "Schema size '{}' is not equal to arrow field size '{}'.",
+                        fieldVectors.size(),
+                        schema.size());
+                throw new DorisException(
+                        "Load Doris data failed, schema size of fetch data is wrong.");
+            }
+            if (fieldVectors.isEmpty() || root.getRowCount() == 0) {
+                logger.debug("One batch in arrow has no data.");
+                return null;
+            }
+            rowCountInOneBatch = root.getRowCount();
+            for (int i = 0; i < rowCountInOneBatch; ++i) {
+                rowBatch.add(new RowBatch.Row(fieldVectors.size()));
+            }
+            convertArrowToRowBatch();
+            readRowCount += root.getRowCount();
+            return this;
+        } catch (DorisException e) {
+            logger.error("Read Doris Data failed because: ", e);
+            throw new DorisRuntimeException(e.getMessage());
+        } catch (IOException e) {
+            return this;
         }
     }
 
@@ -217,24 +258,27 @@ public class RowBatch {
                     return false;
                 }
                 TinyIntVector tinyIntVector = (TinyIntVector) fieldVector;
-                fieldValue = tinyIntVector.isNull(rowIndex) ? null : tinyIntVector.get(rowIndex);
-                addValueToRow(rowIndex, fieldValue);
+                Object fieldValueTiny =
+                        tinyIntVector.isNull(rowIndex) ? null : tinyIntVector.get(rowIndex);
+                addValueToRow(rowIndex, fieldValueTiny);
                 break;
             case "SMALLINT":
                 if (!minorType.equals(Types.MinorType.SMALLINT)) {
                     return false;
                 }
                 SmallIntVector smallIntVector = (SmallIntVector) fieldVector;
-                fieldValue = smallIntVector.isNull(rowIndex) ? null : smallIntVector.get(rowIndex);
-                addValueToRow(rowIndex, fieldValue);
+                Object fieldValueSmallint =
+                        smallIntVector.isNull(rowIndex) ? null : smallIntVector.get(rowIndex);
+                addValueToRow(rowIndex, fieldValueSmallint);
                 break;
             case "INT":
                 if (!minorType.equals(Types.MinorType.INT)) {
                     return false;
                 }
                 IntVector intVector = (IntVector) fieldVector;
-                fieldValue = intVector.isNull(rowIndex) ? null : intVector.get(rowIndex);
-                addValueToRow(rowIndex, fieldValue);
+                Object fieldValueIntVector =
+                        intVector.isNull(rowIndex) ? null : intVector.get(rowIndex);
+                addValueToRow(rowIndex, fieldValueIntVector);
                 break;
             case "IPV4":
                 if (!minorType.equals(Types.MinorType.UINT4)
@@ -248,27 +292,29 @@ public class RowBatch {
                 } else {
                     ipv4Vector = (UInt4Vector) fieldVector;
                 }
-                fieldValue =
+                Object fieldValueIpv4Vector =
                         ipv4Vector.isNull(rowIndex)
                                 ? null
                                 : convertLongToIPv4Address(ipv4Vector.getValueAsLong(rowIndex));
-                addValueToRow(rowIndex, fieldValue);
+                addValueToRow(rowIndex, fieldValueIpv4Vector);
                 break;
             case "BIGINT":
                 if (!minorType.equals(Types.MinorType.BIGINT)) {
                     return false;
                 }
                 BigIntVector bigIntVector = (BigIntVector) fieldVector;
-                fieldValue = bigIntVector.isNull(rowIndex) ? null : bigIntVector.get(rowIndex);
-                addValueToRow(rowIndex, fieldValue);
+                Object fieldValueBigIntVector =
+                        bigIntVector.isNull(rowIndex) ? null : bigIntVector.get(rowIndex);
+                addValueToRow(rowIndex, fieldValueBigIntVector);
                 break;
             case "FLOAT":
                 if (!minorType.equals(Types.MinorType.FLOAT4)) {
                     return false;
                 }
                 Float4Vector float4Vector = (Float4Vector) fieldVector;
-                fieldValue = float4Vector.isNull(rowIndex) ? null : float4Vector.get(rowIndex);
-                addValueToRow(rowIndex, fieldValue);
+                Object fieldValueFloat4Vector =
+                        float4Vector.isNull(rowIndex) ? null : float4Vector.get(rowIndex);
+                addValueToRow(rowIndex, fieldValueFloat4Vector);
                 break;
             case "TIME":
             case "DOUBLE":
@@ -276,23 +322,25 @@ public class RowBatch {
                     return false;
                 }
                 Float8Vector float8Vector = (Float8Vector) fieldVector;
-                fieldValue = float8Vector.isNull(rowIndex) ? null : float8Vector.get(rowIndex);
-                addValueToRow(rowIndex, fieldValue);
+                Object fieldValueFloat8Vector =
+                        float8Vector.isNull(rowIndex) ? null : float8Vector.get(rowIndex);
+                addValueToRow(rowIndex, fieldValueFloat8Vector);
                 break;
             case "BINARY":
                 if (!minorType.equals(Types.MinorType.VARBINARY)) {
                     return false;
                 }
                 VarBinaryVector varBinaryVector = (VarBinaryVector) fieldVector;
-                fieldValue =
+                Object fieldValueVarBinaryVector =
                         varBinaryVector.isNull(rowIndex) ? null : varBinaryVector.get(rowIndex);
-                addValueToRow(rowIndex, fieldValue);
+                addValueToRow(rowIndex, fieldValueVarBinaryVector);
                 break;
             case "DECIMAL":
             case "DECIMALV2":
             case "DECIMAL32":
             case "DECIMAL64":
             case "DECIMAL128I":
+            case "DECIMAL128":
                 if (!minorType.equals(Types.MinorType.DECIMAL)) {
                     return false;
                 }
@@ -316,10 +364,8 @@ public class RowBatch {
                         addValueToRow(rowIndex, null);
                         break;
                     }
-                    String stringValue = new String(date.get(rowIndex));
-                    LocalDate localDate = LocalDate.parse(stringValue, dateFormatter);
-                    //                    LocalDate localDate =
-                    // FastDateUtil.fastParseDate(stringValue,dateTimeV2Formatter.toString());
+                    String stringValue = new String(date.get(rowIndex), StandardCharsets.UTF_8);
+                    LocalDate localDate = FastDateUtil.fastParseDate(stringValue, "yyyy-MM-dd");
                     addValueToRow(rowIndex, localDate);
                 } else {
                     DateDayVector date = (DateDayVector) fieldVector;
@@ -332,28 +378,9 @@ public class RowBatch {
                 }
                 break;
             case "DATETIME":
-                if (!minorType.equals(Types.MinorType.TIMESTAMPMICRO)
-                        && !minorType.equals(Types.MinorType.VARCHAR)) {
-                    return false;
-                }
-                if (minorType.equals(Types.MinorType.VARCHAR)) {
-                    VarCharVector varCharVector = (VarCharVector) fieldVector;
-                    if (varCharVector.isNull(rowIndex)) {
-                        addValueToRow(rowIndex, null);
-                        break;
-                    }
-                    String stringValue = new String(varCharVector.get(rowIndex));
-                    LocalDateTime parse = LocalDateTime.parse(stringValue, dateTimeFormatter);
-                    //                    LocalDateTime parse =
-                    // FastDateUtil.fastParseDateTime(stringValue,dateTimeFormatter.toString());
-                    addValueToRow(rowIndex, parse);
-                } else {
-                    LocalDateTime dateTime = getDateTime(rowIndex, fieldVector);
-                    addValueToRow(rowIndex, dateTime);
-                }
-                break;
             case "DATETIMEV2":
                 if (!minorType.equals(Types.MinorType.TIMESTAMPMICRO)
+                        && !minorType.equals(Types.MinorType.TIMESTAMPSEC)
                         && !minorType.equals(Types.MinorType.VARCHAR)) {
                     return false;
                 }
@@ -363,11 +390,12 @@ public class RowBatch {
                         addValueToRow(rowIndex, null);
                         break;
                     }
-                    String stringValue = new String(varCharVector.get(rowIndex));
+                    String stringValue =
+                            new String(varCharVector.get(rowIndex), StandardCharsets.UTF_8);
                     stringValue = completeMilliseconds(stringValue);
-                    LocalDateTime parse = LocalDateTime.parse(stringValue, dateTimeV2Formatter);
-                    //                    LocalDateTime parse =
-                    // FastDateUtil.fastParseDateTime(stringValue,dateTimeV2Formatter.toString());
+                    LocalDateTime parse =
+                            FastDateUtil.fastParseDateTime(
+                                    stringValue, "yyyy-MM-dd HH:mm:ss.SSSSSS");
                     addValueToRow(rowIndex, parse);
                 } else {
                     LocalDateTime dateTime = getDateTime(rowIndex, fieldVector);
@@ -403,7 +431,8 @@ public class RowBatch {
                         addValueToRow(rowIndex, null);
                         break;
                     }
-                    String stringValue = new String(largeIntVector.get(rowIndex));
+                    String stringValue =
+                            new String(largeIntVector.get(rowIndex), StandardCharsets.UTF_8);
                     BigInteger largeInt = new BigInteger(stringValue);
                     addValueToRow(rowIndex, largeInt);
                     break;
@@ -421,7 +450,8 @@ public class RowBatch {
                     addValueToRow(rowIndex, null);
                     break;
                 }
-                String stringValue = new String(varCharVector.get(rowIndex));
+                String stringValue =
+                        new String(varCharVector.get(rowIndex), StandardCharsets.UTF_8);
                 addValueToRow(rowIndex, stringValue);
                 break;
             case "IPV6":
@@ -433,7 +463,8 @@ public class RowBatch {
                     addValueToRow(rowIndex, null);
                     break;
                 }
-                String ipv6Str = new String(ipv6VarcharVector.get(rowIndex));
+                String ipv6Str =
+                        new String(ipv6VarcharVector.get(rowIndex), StandardCharsets.UTF_8);
                 String ipv6Address = IPUtils.fromBigInteger(new BigInteger(ipv6Str));
                 addValueToRow(rowIndex, ipv6Address);
                 break;
@@ -486,7 +517,33 @@ public class RowBatch {
 
     @VisibleForTesting
     public LocalDateTime getDateTime(int rowIndex, FieldVector fieldVector) {
+        if (fieldVector instanceof TimeStampMicroVector) {
+            return getDateTimeFromTimeStampMicro(rowIndex, fieldVector);
+        } else {
+            return getDateTimeFromTimeStampSec(rowIndex, fieldVector);
+        }
+    }
+
+    private LocalDateTime getDateTimeFromTimeStampMicro(int rowIndex, FieldVector fieldVector) {
         TimeStampMicroVector vector = (TimeStampMicroVector) fieldVector;
+        if (vector.isNull(rowIndex)) {
+            return null;
+        }
+        long time = vector.get(rowIndex);
+        Instant instant;
+        if (time / 10000000000L == 0) { // datetime(0)
+            instant = Instant.ofEpochSecond(time);
+        } else if (time / 10000000000000L == 0) { // datetime(3)
+            instant = Instant.ofEpochMilli(time);
+        } else { // datetime(6)
+            instant = Instant.ofEpochSecond(time / 1000000, time % 1000000 * 1000);
+        }
+        LocalDateTime dateTime = LocalDateTime.ofInstant(instant, ZoneId.systemDefault());
+        return dateTime;
+    }
+
+    private LocalDateTime getDateTimeFromTimeStampSec(int rowIndex, FieldVector fieldVector) {
+        TimeStampSecVector vector = (TimeStampSecVector) fieldVector;
         if (vector.isNull(rowIndex)) {
             return null;
         }
@@ -513,14 +570,30 @@ public class RowBatch {
             return stringValue;
         }
 
-        StringBuilder sb = new StringBuilder(stringValue);
         if (stringValue.length() == DATETIME_PATTERN.length()) {
-            sb.append(".");
+            stringValue += ".";
         }
-        while (sb.toString().length() < DATETIMEV2_PATTERN.length()) {
-            sb.append(0);
+        int s = DATETIMEV2_PATTERN.length() - stringValue.length();
+        switch (s) {
+            case 1:
+                return stringValue + "0";
+            case 2:
+                return stringValue + "00";
+            case 3:
+                return stringValue + "000";
+            case 4:
+                return stringValue + "0000";
+            case 5:
+                return stringValue + "00000";
+            case 6:
+                return stringValue + "000000";
+            default:
+                return stringValue;
         }
-        return sb.toString();
+        //        while (sb.toString().length() < DATETIMEV2_PATTERN.length()) {
+        //            sb.add("0");
+        //        }
+        //        return sb.toString();
     }
 
     public List<Object> next() {
