+++ flink-doris-connector/src/main/java/org/apache/doris/flink/source/reader/ValueReader.java
@@ -1,0 +1,11 @@
+package org.apache.doris.flink.source.reader;
+
+import java.util.List;
+
+public interface ValueReader {
+    boolean hasNext();
+
+    List next();
+
+    void close() throws Exception;
+}
